v1.0  (12/8/15)
This program will analyze a year's precipitation data for a fictional location. The user is tasked with entering the amount of
rainfall (in inches) for each month of the year starting with month 1 (Jan) up to month 12 (Dec). The program will calculate the total
rainfall for the year, the average monthly rainfall, determine the driest month, and determine the wettest month. Finally, the program
will display the data in a comprehensive report.